change eventListener callback function
For the starBtn, fireBtn, and resetSalesBtn the callback function was changed.  Now the evenListener calls a named function.  The named function contains the code that was previously held in the callback function of the event listener.
In the resetSales code, there was a bug with the achievementsEl.textContent line.  This line was changed to achievementsCount.textContent.